feat: change trades from int slice to float64 slice

+++ candlestick/candlestick_test.go
@@ -7,7 +7,7 @@ import (
 )
 
 func TestAggregateTrades(t *testing.T) {
-	trades := []int{14, 14, 15, 15, 16, 15, 13, 12, 13}
+	trades := []float64{14, 14, 15, 15, 16, 15, 13, 12, 13}
 	want := candlestick.TradeAggregate{Open: 14, Close: 13, High: 16, Low: 12}
 	got := candlestick.AggregateTrades(trades)
 	assert.Equal(t, want.Open, got.Open, "Open")
@@ -17,14 +17,14 @@ func TestAggregateTrades(t *testing.T) {
 }
 
 func BenchmarkAggregateTrades(b *testing.B) {
-	trades := []int{14, 14, 15, 15, 16, 15, 13, 12, 14, 15, 15, 16, 15, 13, 12, 14, 15, 15, 16, 15, 13, 12, 14, 15, 15, 16, 15, 13, 12, 14, 15, 15, 16, 15, 13, 12, 13}
+	trades := []float64{14, 14, 15, 15, 16, 15, 13, 12, 14, 15, 15, 16, 15, 13, 12, 14, 15, 15, 16, 15, 13, 12, 14, 15, 15, 16, 15, 13, 12, 14, 15, 15, 16, 15, 13, 12, 13}
 	for n := 0; n < b.N; n++ {
 		candlestick.AggregateTrades(trades)
 	}
 }
 
 func BenchmarkAggregateTradesSideEffects(b *testing.B) {
-	trades := []int{14, 14, 15, 15, 16, 15, 13, 12, 14, 15, 15, 16, 15, 13, 12, 14, 15, 15, 16, 15, 13, 12, 14, 15, 15, 16, 15, 13, 12, 14, 15, 15, 16, 15, 13, 12, 13}
+	trades := []float64{14, 14, 15, 15, 16, 15, 13, 12, 14, 15, 15, 16, 15, 13, 12, 14, 15, 15, 16, 15, 13, 12, 14, 15, 15, 16, 15, 13, 12, 14, 15, 15, 16, 15, 13, 12, 13}
 	for n := 0; n < b.N; n++ {
 		candlestick.AggregateTradesSideEffects(trades)
 	}
